Making DailyTaskInstance be an actual Node

+++ backend/src/task-completion/task-completion-mutations.ts
@@ -7,6 +7,8 @@ import type { MutationResolvers } from '../graphql/resolver-types.ts'
 import { assertAuthenticated, type Context } from '../graphql/context.ts'
 import { fromGlobalId } from '../globalId.ts'
 import { GraphQLError } from 'graphql'
+import { tableToDomain as routineSlotTableToDomain } from '../routine-slot/routine-slot-domain.ts'
+import { dailyTaskInstanceToGraphQL } from '../schedule/schedule-domain.ts'
 
 export const completeRoutineSlot: MutationResolvers<Context>['completeRoutineSlot'] =
   async (_parent, { routineSlotId }, context) => {
@@ -14,11 +16,11 @@ export const completeRoutineSlot: MutationResolvers<Context>['completeRoutineSlo
 
     const decodedRoutineSlotId = fromGlobalId(routineSlotId, 'RoutineSlot')
 
-    const routineSlot = await context.routineRepo.findByIdAndUserId(
+    const routineSlotRow = await context.routineRepo.findByIdAndUserId(
       decodedRoutineSlotId,
       context.currentUser.id,
     )
-    if (!routineSlot) {
+    if (!routineSlotRow) {
       throw new GraphQLError('Routine slot not found')
     }
 
@@ -29,8 +31,14 @@ export const completeRoutineSlot: MutationResolvers<Context>['completeRoutineSlo
 
     const completion = tableToDomain(completionRow)
     const completionEdge = buildTaskCompletionEdge(completion)
+    const routineSlot = routineSlotTableToDomain(routineSlotRow)
 
     return {
+      dailyTaskInstance: dailyTaskInstanceToGraphQL({
+        date: completion.completedAt,
+        routineSlot,
+        completion,
+      }),
       taskCompletionEdge: {
         node: taskCompletionToGraphQL(completionEdge.node),
         cursor: completionEdge.cursor,
@@ -43,6 +51,24 @@ export const uncompleteRoutineSlot: MutationResolvers<Context>['uncompleteRoutin
     assertAuthenticated(context)
 
     const id = fromGlobalId(taskCompletionId, 'TaskCompletion')
+
+    const existingRow = await context.taskCompletionRepo.findByIdAndUserId(
+      id,
+      context.currentUser.id,
+    )
+    if (!existingRow) {
+      throw new GraphQLError("Task completion didn't exist")
+    }
+    const existing = tableToDomain(existingRow)
+
+    const routineSlotRow = await context.routineRepo.findByIdAndUserId(
+      existing.routineSlotId,
+      context.currentUser.id,
+    )
+    if (!routineSlotRow) {
+      throw new GraphQLError('Routine slot not found')
+    }
+
     const res = await context.taskCompletionRepo.deleteCompletion(
       id,
       context.currentUser.id,
@@ -53,6 +79,11 @@ export const uncompleteRoutineSlot: MutationResolvers<Context>['uncompleteRoutin
     }
 
     return {
+      dailyTaskInstance: dailyTaskInstanceToGraphQL({
+        date: existing.completedAt,
+        routineSlot: routineSlotTableToDomain(routineSlotRow),
+        completion: null,
+      }),
       deletedId: taskCompletionId,
     }
   }
